Modulize component for home page
