+++ website/pages/en/index.js
@@ -7,116 +7,49 @@
 
 const React = require('react');
 
+// Built-in library
 const CompLibrary = require('../../core/CompLibrary.js');
-const MarkdownBlock = CompLibrary.MarkdownBlock; /* Used to read markdown */
 const Container = CompLibrary.Container;
 const GridBlock = CompLibrary.GridBlock;
 
-const siteConfig = require(process.cwd() + '/siteConfig.js');
-
-function imgUrl(img) {
-  return siteConfig.baseUrl + 'img/' + img;
-}
-
-function docUrl(doc, language) {
-  return siteConfig.baseUrl + 'docs/' + (language ? language + '/' : '') + doc;
-}
-
-function pageUrl(page, language) {
-  return siteConfig.baseUrl + (language ? language + '/' : '') + page;
-}
-
-class Button extends React.Component {
-  render() {
-    return (
-      <div className="pluginWrapper buttonWrapper">
-        <a className="button" href={this.props.href} target={this.props.target}>
-          {this.props.children}
-        </a>
-      </div>
-    );
-  }
-}
-
-Button.defaultProps = {
-  target: '_self',
-};
-
-const SplashContainer = props => (
-  <div className="homeContainer">
-    <div className="homeSplashFade">
-      <div className="wrapper homeWrapper">{props.children}</div>
-    </div>
-  </div>
-);
-
-const Logo = props => (
-  <div className="projectLogo">
-    <img src={props.img_src} />
-  </div>
-);
-
-const ProjectTitle = props => (
-  <h2 className="projectTitle">
-    {siteConfig.title}
-    <small>{siteConfig.tagline}</small>
-  </h2>
-);
-
-const PromoSection = props => (
-  <div className="section promoSection">
-    <div className="promoRow">
-      <div className="pluginRowBlock">{props.children}</div>
-    </div>
-  </div>
-);
-
-class HomeSplash extends React.Component {
-  render() {
-    let language = this.props.language || '';
-    return (
-      <SplashContainer>
-        {/* <Logo img_src={imgUrl('docusaurus.svg')} /> */}
-        <div className="inner">
-          <ProjectTitle />
-          <PromoSection>
-            <Button href={docUrl('getting-started.html', language)}>Try It Out</Button>
-            <Button href='./storybook/index.html'>Live Demo</Button>
-            <Button href={docUrl('table-props.html', language)}>Docs</Button>
-          </PromoSection>
-        </div>
-      </SplashContainer>
-    );
-  }
-}
+// Customized component and utils
+const HomeSplash = require(process.cwd() + '/core/homeSplash.js');
+const utils = require(process.cwd() + '/core/utils');
+const imgUrl = utils.imgUrl;
 
 const Block = props => (
   <Container
     padding={['bottom', 'top']}
     id={props.id}
-    background={props.background}>
-    <GridBlock align={ props.align || 'center' } contents={props.children} layout={props.layout} />
+    background={props.background}
+  >
+    <GridBlock
+      align={ props.align || 'center' }
+      layout={props.layout}
+      className={props.className}
+      contents={props.children}
+    />
   </Container>
 );
 
-const Features = props => (
-  <Block layout="fourColumn">
+const Features = () => (
+  <Block layout="threeColumn" className="feature-block">
     {[
       {
         content: 'Sortable, Row Selection, Cell Editor, Row Expand, Column Filter Pagination etc.',
-        // image: imgUrl('docusaurus.svg'),
+        image: imgUrl('icon/bulb.svg'),
         imageAlign: 'top',
         title: 'Rich Functionality',
       },
       {
         content: 'Configurable and customizable table',
-        // image: imgUrl('docusaurus.svg'),
+        image: imgUrl('icon/tool.svg'),
         imageAlign: 'top',
         title: 'Customization',
       },
       {
         content: 'Satisfy for Redux/Mobx or any other state management tool.',
-        // image: imgUrl('docusaurus.svg'),
+        image: imgUrl('icon/store.svg'),
         imageAlign: 'top',
         title: 'Remote',
       },
@@ -124,20 +57,11 @@ const Features = props => (
   </Block>
 );
 
-// const FeatureCallout = props => (
-//   <div
-//     className="productShowcaseSection paddingBottom"
-//     style={{textAlign: 'center'}}>
-//     <h2>Feature Callout</h2>
-//     <MarkdownBlock>These are features of this project</MarkdownBlock>
-//   </div>
-// );
-
-const LearnHow = props => (
+const LearnHow = () => (
   <Block background="light" align="left">
     {[
       {
-        content: 'Intuitive to use. <br/>Compatitable for Bootstrap 3 and 4. <br/>Better than legacy react-bootstrap-table2!!<br/>',
+        content: 'Intuitive to use. <br/>Compatible for Bootstrap 3 and 4. <br/>Better than legacy react-bootstrap-table2!!<br/>',
         image: imgUrl('react-bootstrap-table2-sample.png'),
         imageAlign: 'right',
         title: 'react-bootstrap-table2',
@@ -145,62 +69,6 @@ const LearnHow = props => (
     ]}
   </Block>
 );
-
-const TryOut = props => (
-  <Block id="try">
-    {[
-      {
-        content: 'Talk about trying this out',
-        image: imgUrl('docusaurus.svg'),
-        imageAlign: 'left',
-        title: 'Try it Out',
-      },
-    ]}
-  </Block>
-);
-
-const Description = props => (
-  <Block background="dark">
-    {[
-      {
-        content: 'This is another description of how this project is useful',
-        image: imgUrl('docusaurus.svg'),
-        imageAlign: 'right',
-        title: 'Description',
-      },
-    ]}
-  </Block>
-);
-
-const Showcase = props => {
-  if ((siteConfig.users || []).length === 0) {
-    return null;
-  }
-  const showcase = siteConfig.users
-    .filter(user => {
-      return user.pinned;
-    })
-    .map((user, i) => {
-      return (
-        <a href={user.infoLink} key={i}>
-          <img src={user.image} title={user.caption} />
-        </a>
-      );
-    });
-
-  return (
-    <div className="productShowcaseSection paddingBottom">
-      <h2>{"Who's Using This?"}</h2>
-      <p>This project is used by all these people</p>
-      <div className="logos">{showcase}</div>
-      <div className="more-users">
-        <a className="button" href={pageUrl('users.html', props.language)}>
-          More {siteConfig.title} Users
-        </a>
-      </div>
-    </div>
-  );
-};
 
 class Index extends React.Component {
   render() {
@@ -211,11 +79,7 @@ class Index extends React.Component {
         <HomeSplash language={language} />
         <div className="mainContainer">
           <Features />
-          {/* <FeatureCallout /> */}
           <LearnHow />
-          {/* <TryOut />
-          <Description />
-          <Showcase language={language} /> */}
         </div>
       </div>
     );
