+++ website/static/css/custom.css
@@ -14,3 +14,7 @@
 
 @media only screen and (min-width: 1500px) {
 }
+
+.feature-block .blockImage img {
+  width: 48px;
+}
